refactor: improve namings
this is mostly just UI-facing; need to follow-up with updates in code
once we know these changes are desirable.

embodies -> fulfills
claims -> justification
TopicToolbar -> WorkspaceToolbar
structure -> breakdown
